desktop_app/PythonProxy_v2/requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
click-spinner==0.1.10
colorama==0.4.6
croniter==1.4.1
cryptography==41.0.7
fastapi==0.104.1
fastapi-utilities==0.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.6
jwt==1.3.1
kombu==5.3.4
mysql-connector-python==8.2.0
prompt-toolkit==3.0.43
protobuf==4.21.12
pycparser==2.21
pydantic==2.5.2
pydantic-core==2.14.5
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
pyyaml==6.0.1
redis==5.0.1
schedule==1.2.1
setuptools==68.0.0
six==1.16.0
sniffio==1.3.0
sqlalchemy==2.0.24
starlette==0.27.0
typing-extensions==4.8.0
uvicorn==0.24.0.post1
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.12
websockets==12.0
wheel==0.41.2
azure-storage-blob=12.19.1